Move IPv4 and IPv6 management into dedicated templates
The IPv4 and IPv6 management code is duplicated between
vlan-management.rsc and vlans-custom.rsc. To reduce this and improve
IPv6 handling across both cases, move the IPv4 and IPv6 code into
dedicated files which can then be called by the parent template with the
VLAN passed into the sub-template as the global variable.