Implement support for managing SSL certificates
Implement support for managing SSL certificates for both the network
(i.e. private CA certificates) and the host (host SSL certificate), and
configure the WebFig and API services running over SSL with the host SSL
certificate for remote access.

This is combined with the SSH key setup as they are generally managed on
initial installation and rarely configured.