Add edge and point-to-point configuration for bridge ports
Add the edge configuration for bridge ports, which will help speed up
connections on ports not connected to other bridges (i.e. switched) or
trunking VLANs, and hopefully trunking VLANs too by enabling default
options with discovery.

Also set/reset point-to-point to the default of auto in case this
setting it changed, which again should help manage the speed up of
connections to all types of ports.