Fix error in the blocked vlan untagged ports processing
Fix the error in the blocked vlan untagged ports processing, which
didn't have a distinction between normal and netinstall cases, and as
such was always setting blocked as an untagged port for anything not
specifically set to management regardless.

This will ensure that outside of netinstall exports, interfaces will
only be added to the blocked VLAN when it is explicitly configured for
it.