Update services configuration to prevent disabling
Update the services configuration for www-ssl and api-ssl support to
ensure that there is a check for the certificate first, and then ensure
the configuration is set to use that certificate, otherwise it's
disabled.

This should help when running update.rsc after certificates.rsc has been
run, allowing the service to remain configured rather than reset.