Change gateway handling for IPv4 and IPv6 to be network set
Change the handling around IPv4 and IPv6 so that the value is configured
in the network configuration as a "global" value, rather than a specific
value per each host under it's management VLAN settings.

Also allow it to be overridden under IPv4, with the default being the
VLAN address used under the VLAN configuration.